mgmt, update sql to 2022-11-01-preview (#35911)
mgmt, local generation for sql

+++ sdk/resourcemanager/azure-resourcemanager-sql/src/main/java/com/azure/resourcemanager/sql/fluent/ManagedDatabaseSecurityEventsClient.java
@@ -35,8 +35,8 @@ public interface ManagedDatabaseSecurityEventsClient {
         String managedInstanceName,
         String databaseName,
         String filter,
-        Long skip,
-        Long top,
+        Integer skip,
+        Integer top,
         String skiptoken);
 
     /**
@@ -94,8 +94,8 @@ public interface ManagedDatabaseSecurityEventsClient {
         String managedInstanceName,
         String databaseName,
         String filter,
-        Long skip,
-        Long top,
+        Integer skip,
+        Integer top,
         String skiptoken,
         Context context);
 }
